Style changes to conform to the Awesome List guidelines.
Among the Awesome List style guide recommendations is to omit the
leading "A" or "An" preposition for line item descriptions. This commit
removes all leading "A/An" words from line items, so as to better
conform to the Awesome List style guide.

This commit further fixes several typos ("opensource" becomes "open
source"), and capitalization errors (title case line items in some
instances are normalized to standard initial-caps on the first word,
which is no longer "A" or "An").

Additionally, this commit replaces the ampersand (`&`) in headlines and
line item descriptions with its English equivalent, "and," in order to
read more naturally. This change also helps resolve some ambiguity in
URL fragments (intra-page anchors), which would otherwise contain two
dashes in sequence (`--`) rather than a semantic description of the
section to which the intra-page link target actually refers.